Refactor reservation component structure
Moved reservation components into more context-appropriate subdirectories, improving project structure and clarity. Adjusted imports accordingly to reflect these changes.
