+++ src/app/reservation/complete/components/ReservationDetails.tsx
