+++ src/app/reservation/confirm/components/NotesSection.tsx
@@ -2,12 +2,12 @@
 
 import React from "react";
 import { AlertCircle } from "lucide-react";
-import { IconWrapper } from "./IconWrapper";
+import { IconWrapper } from "@/components/shared/IconWrapper";
 
 export const NotesSection = () => {
   return (
     <div className="rounded-lg p-4 mb-6">
-      <h3 className="text-lg font-medium mb-4">留意事項</h3>
+      <h3 className="text-2xl font-bold mb-6">注意事項</h3>
       <div className="space-y-4">
         <div className="flex items-center gap-3">
           <IconWrapper color="warning">
