+++ src/app/reservation/confirm/components/OpportunityInfo.tsx
