+++ src/app/reservation/confirm/components/PaymentSection.tsx
